+++ components/Item/components/Handle/Handle.tsx
@@ -1,6 +1,6 @@
-import React, {forwardRef} from 'react';
+import React, { forwardRef } from 'react';
 
-import {Action, ActionProps} from '../Action';
+import { Action, ActionProps } from '../Action';
 
 export const Handle = forwardRef<HTMLButtonElement, ActionProps>(
   (props, ref) => {
@@ -18,3 +18,5 @@ export const Handle = forwardRef<HTMLButtonElement, ActionProps>(
     );
   }
 );
+
+Handle.displayName = 'Handle';
