+++ package.json
@@ -36,6 +36,7 @@
 		"next-auth": "^4.24.5",
 		"next-swagger-doc": "^0.4.0",
 		"next-themes": "^0.2.1",
+		"postcss-url": "^10.1.3",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-grid-layout": "^1.4.4",
